1131_Chatbot_Final/requirements.txt
flask
pillow==11.0.0
line-bot-sdk
google-generativeai
azure-ai-translation-text
azure-cognitiveservices-speech
azure-ai-textanalytics
